Fixed non-bug with Ratnums, flirted with GMP
As usual, the bug was in the most recently written code, which by
convention I checked last! The random numeric generator was really
screwing memory up by passing the same cpp_int around to different
cpp_rationals to be their numerator; each of them freed it upon going
out of scope of course. The ratnum generator is now fixed and working as
intended.

While debugging, I tried swapping out the arbitrary precision backend to
see if that's where the problem was. I tried GMP and while I was at it,
I tested its speed once the generator was fixed (trying GMP was a
project goal). Unfortunately, it doesn't seem to be much faster than the
cpp_int backend so I reversed course to avoid another dependency.