MANY changes (bugfix for numerics, python testing base)
TODO: Have the benchmarks print out flat data and do the analysis on the
data later rather than as they go. This will allow the data to be used
in ways that are unforeseen as of now.

Added a SWIG wrapper for the library so that it can be unit tested from
python. I think this will be much easier than taking on another external
dependency (I don't view python as a dependency since it's so widely
used and available). Adding this wrapper exposed the bug immediately
below this paragraph, as the wrapper was segfaulting.

Numeric and NumericInternal did not have their move/copy constructors
deleted, so they were making copies I didn't know about and moving
internals without cleaning up the source (preventing it from freeing the
pointer). To prevent this, move/copy constructors were implemented
properly for all Numerics.

Fixed the benchmark_bignum_fixnum_int64_t comparison and added it to the
driver. Added a mini-driver for it as well for solo runs.

Added a benchmark to measure the overhead resulting from the Numeric
RAII shell around ratnums, since ratnums appear to run so much more
slowly than bignums. Turns out the overhead is minimal. Still
investigating the time sink -- I would expect GCD simplifying to occur
at the end, once, with a cost of log n (n is the # of bit chunks), so I
don't believe that is the cause.

Added operation symmetry benchmark to check for any issues where
left_operand_type OP right_operand_type is taking significantly longer
than right_operand_type OP left_operand_type for operands of the same
size. This would indicate a bug unless otherwise explainable. It has
already highlighted a peculiarity where division is taking much longer
than I think it should in some cases.

Adjusted the benchmark settings and definitions some more, still
tweaking what should be adjustable. On a related note I made some
modifications to the number of loops for some of the benchmarks. They
were a bit unbearably long-running before.