Leveraging python testing, huge bugfix patch
Python testing is continuing to be expanded (not all operations added,
not all added operations passing).

Many, many operations were incorrect and needed to be modified. This is to be expected since real correctness testing hasn't been done since the ground-up refactor.

This has changed their performance characteristics so they are being
re-benchmarked on CADE.

Committing in preparation for deep modifications to the codebase, need a
save point.